LPS-73770 We don't need to rebuild the whole tree; we can just update the treePath of each group whose parentGroupId we modify
We can do this because the following two conditions are both true:

1. Each parentGroupId value that we are setting corresponds to a group that already has the correct treePath.
2. Each group we are modifying the parentGroupId of cannot have any children groups that wouldn't also get touched by this fix.

To elaborate on condition (1), we are always going to modify a staging group to have a parentGroupId of a live group. It is impossible for a live group to have any staging groups in its ancestry. Since we are only modifying staging groups in this fix, any live groups will not need to have their treePaths modified during this fix. So we can assume that the treePath of any live group is already valid even while this code is running.

To elaborate on condition (2), it is impossible for any staging group to have a live groups as a descendant. I also found in my testing that it seemed to be impossible for a local staging group to have a remote staging group as a descendant. Since the only type of groups we are not modifying are remote staging groups and live groups, we can safely assume that we will not break the treePath of any descendant of the groups we are modifying.