LPS-74160 Modify onAfterUpdate methods to throw the Exception if the LDAP user already exists
The logic here is that if the LDAP user already exists, we will only cath an exception if we are modifying a field to a value that LDAP is unhappy with. This is true because the way LDAPUserExporter works is it only tries to modify the fields that it sees are different from the fields that already exist in LDAP. If we are trying to modify a field to a value that LDAP is unhappy with, we should throw an exception to prevent inconsistencies between the database and LDAP.

If, however, the user does not exist in LDAP, then that means there is already an inconsistency between LDAP and the database anyway for this particular user. In this case, we do not know whether or not we are modifying a field to a value that LDAP is unhappy with, or whether that field already had that value. It doesn't matter either way as we do not expect this user to end up in LDAP. This should only ever be possible for users who existed before LDAP was configured