LPS-74160 The logic for onAfterAdd/RemoveAssociation is similar to that of onAfterUpdate
If the user does not exist in LDAP, there is no harm in allowing him/her to join or unjoin a User Group since LDAP cannot keep track of that anyway. However, if the user does exist in LDAP, then we must make sure that whether or not LDAP thinks the user is part of a group is consistent with what the database says.